Refactor rights declaration
